1. 전사 캐릭터에 칼 추가 2. warrior_0 스킬 작업중(애니메이션 관련)
애니메이션이 동작할 때 좀 더 자연스럽게 되거나 움직임을 캔슬하는 모습 등이 필요할듯함. 점프할 때도 마찬가지고, 스킬 애니메이션 끝날때 이벤트 발생시켜서 콜백하는것도 필요해보이고 콤보어택같은 것도 필요할듯. 애니메이션에 대해 전반적으로 심화된 공부가 필요함. 02. 12. ~

+++ Assets/00_Embers/02_Scripts/Core/Enemy/Enemy.cs
@@ -5,8 +5,8 @@ namespace NOLDA
 {
     public class Enemy : NetworkBehaviour
     {
-        [SyncVar] public int maxHp = 100;
-        [SyncVar] private int currentHp;
+        [SyncVar] public float maxHp = 100;
+        [SyncVar] private float currentHp;
 
         private void Start()
         {
@@ -14,7 +14,7 @@ namespace NOLDA
         }
 
         [Server]
-        public void TakeDamage(int damage)
+        public void TakeDamage(float damage)
         {
             currentHp -= damage;
             Debug.Log($"몬스터 {gameObject.name}가 {damage}의 피해를 받음. 현재 HP: {currentHp}");
